skip update if GPUs are unavailable (e.g. ROCm build test)

+++ .pfnci/wheel-linux/update-cuda-driver.sh
@@ -2,7 +2,7 @@
 
 set -ue
 
-if dpkg -s cuda-drivers-495; then
+if dpkg -s cuda-drivers-495 && ls /dev/nvidiactl ; then
     killall Xorg
     nvidia-smi -pm 0
